% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Jags-Ymet-Xnom1fac-MnormalHom.R
\name{genMCMC_ANOVA}
\alias{genMCMC_ANOVA}
\title{Bayesian ANOVA with homogenous variance}
\usage{
genMCMC_ANOVA(datFrm, yName = "y", xName = "x", numSavedSteps = 50000,
  thinSteps = 1, saveName = NULL)
}
\arguments{
\item{datFrm}{}

\item{yName}{}

\item{xName}{}

\item{numSavedSteps}{}

\item{thinSteps}{}

\item{saveName}{}
}
\description{
Bayesian ANOVA with homogenous variance
}

